normed

+++ src/print_sha256.c
@@ -6,7 +6,7 @@
 /*   By: ggregoir <ggregoir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/16 22:33:18 by ggregoir          #+#    #+#             */
-/*   Updated: 2019/07/19 01:59:59 by ggregoir         ###   ########.fr       */
+/*   Updated: 2019/07/25 19:31:18 by ggregoir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 void		print_sha256(char *to_hash)
 {
-	size_t		len;
+	uint32_t	len;
 	t_sha256	h;
 
 	ft_bzero(&h, sizeof(h));
 	len = ft_strlen(to_hash);
 	init_hash_sha256(&h);
 	sha256_algo(to_hash, len, &h);
-	ft_printf("%.8x%.8x%.8x%.8x%.8x%.8x%.8x%.8x",h.h0,h.h1,h.h2,h.h3,
-		h.h4,h.h5,h.h6,h.h7);
+	ft_printf("%.8x%.8x%.8x%.8x%.8x%.8x%.8x%.8x", h.h0, h.h1, h.h2, h.h3,
+		h.h4, h.h5, h.h6, h.h7);
 }
 
 void		sha256_p(char *to_hash, int8_t *flags)
@@ -89,9 +89,9 @@ void		sha256_noflag(char *to_hash, int8_t *flags)
 		filename = to_hash;
 		if ((to_hash = get_file(to_hash)) != NULL)
 		{
-				ft_printf("SHA256(%s)= ", filename);
-				print_sha256(to_hash);
-				ft_printf("\n");
+			ft_printf("SHA256(%s)= ", filename);
+			print_sha256(to_hash);
+			ft_printf("\n");
 		}
 	}
 }
